requirements.txt
pyspark==3.5.1
python-dotenv==1.0.1
psycopg[binary]==3.2.3
